tasks_service/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
kafka-python
